Add explicit casting

+++ Gameplay/Combat/DamageNumber.cpp
@@ -101,7 +101,7 @@ namespace ms
 				for (size_t i = 0; i < restnum.length(); i++)
 				{
 					char c = restnum[i];
-					Point<int16_t> yshift = {0, (i % 2) ? -2 : 2};
+                    Point<int16_t> yshift = {0, static_cast<short>((i % 2) ? -2 : 2)};
 					charsets[type][true].draw(c, {position + yshift, interopc});
 
 					int16_t advance;
