+++ Template/Point.h
@@ -154,19 +154,19 @@ namespace ms
 		// Return a point whose coordinates are the negation of this point's coordinates
 		constexpr Point<T> operator - () const
 		{
-			return { -a, -b };
+            return { static_cast<short>(-a), static_cast<short>(-b) };
 		}
 
 		// Return a point whose coordinates have been added the specified amount
 		constexpr Point<T> operator + (T v) const
 		{
-			return { a + v, b + v };
+            return { static_cast<short>(a + v), static_cast<short>(b + v) };
 		}
 
 		// Return a point whose coordinates have been subtracted the specified amount
 		constexpr Point<T> operator - (T v) const
 		{
-			return { a - v, b - v };
+            return { static_cast<short>(a - v), static_cast<short>(b - v) };
 		}
 
 		// Return a point whose coordinates have been multiplied by the specified amount
@@ -178,19 +178,19 @@ namespace ms
 		// Return a point whose coordinates have been divided by the specified amount
 		constexpr Point<T> operator / (T v) const
 		{
-			return { a / v, b / v };
+            return { static_cast<short>(a / v), static_cast<short>(b / v) };
 		}
 
 		// Return a point whose coordinates are the sum of this and another points coordinates
 		constexpr Point<T> operator + (Point<T> v) const
 		{
-			return { a + v.a, b + v.b };
+            return { static_cast<short>(a + v.a), static_cast<short>(b + v.b) };
 		}
 
 		// Return a point whose coordinates are the difference of this and another points coordinates
 		constexpr Point<T> operator - (Point<T> v) const
 		{
-			return { a - v.a, b - v.b };
+            return { static_cast<short>(a - v.a), static_cast<short>(b - v.b) };
 		}
 
 		// Return a point whose coordinates are the product of this and another points coordinates
@@ -203,8 +203,8 @@ namespace ms
 		constexpr Point<T> operator / (Point<T> v) const
 		{
 			return {
-				a / (v.a == 0 ? 1 : v.a),
-				b / (v.b == 0 ? 1 : v.b)
+                static_cast<short>(a / (v.a == 0 ? 1 : v.a)),
+                static_cast<short>(b / (v.b == 0 ? 1 : v.b))
 			};
 		}
